requirements.txt
PyYAML
keyboard
pyaudio
pydub
openai
python-dotenv
requests
